data_extraction_pipline/requirements.txt
requests

tqdm
python-dotenv
pandas
biopython
beautifulsoup4
pymupdf            # fitz
sentence-transformers
chromadb
spacy
scispacy
regex
transformers
rfc3339
tqdm
pyyaml
